requests: customize permission policy per type
