+++ invenio_requests/config.py
@@ -2,6 +2,7 @@
 #
 # Copyright (C) 2021 CERN.
 # Copyright (C) 2021 - 2022 TU Wien.
+# Copyright (C) 2022 Northwestern University.
 #
 # Invenio-Requests is free software; you can redistribute it and/or modify it
 # under the terms of the MIT License; see LICENSE file for more details.
@@ -10,12 +11,10 @@
 
 from invenio_records_resources.references.resolvers import UserResolver
 
-from invenio_requests.services.requests import facets
+from .services.requests import facets
 
-from .services.permissions import PermissionPolicy
-
-REQUESTS_PERMISSION_POLICY = PermissionPolicy
-"""Override the default requests/comments permission policy."""
+REQUESTS_PERMISSION_POLICY = None
+"""Deprecated. Override the default requests/comments permission policy."""
 
 REQUESTS_REGISTERED_TYPES = []
 """Configuration for registered Request Types."""
