+++ invenio_requests/customizations/base/request_types.py
@@ -2,6 +2,7 @@
 #
 # Copyright (C) 2021 - 2022 TU Wien.
 # Copyright (C) 2021 CERN.
+# Copyright (C) 2022 Northwestern University.
 #
 # Invenio-Requests is free software; you can redistribute it and/or modify
 # it under the terms of the MIT License; see LICENSE file for more details.
@@ -19,6 +20,7 @@ import marshmallow as ma
 from invenio_records_resources.services.references import EntityReferenceBaseSchema
 
 from ...proxies import current_requests
+from .permissions import BaseRequestPermissionPolicy
 
 
 class RequestType:
@@ -95,6 +97,8 @@ class RequestType:
             # ...
         }
     """
+
+    permission_policy_cls = BaseRequestPermissionPolicy
 
     @classmethod
     def _create_marshmallow_schema(cls):
